hmr final touch
