+++ config/webpack.server.config.babel.js
@@ -4,7 +4,6 @@ const baseConfig = require('./webpack.base.config')
 const project = require('./project.config')
 const { assetUrl } = require('./env.config').default
 const path = require('path')
-const StartServerPlugin = require('start-server-webpack-plugin')
 
 let config = {
   target: 'node',
@@ -39,11 +38,7 @@ let config = {
 
 if (project.globals.__DEV__) {
   const addConfig = {
-    plugins: [
-      new StartServerPlugin({
-        name: 'bundle.js'
-      })
-    ]
+    plugins: []
   }
   config = merge(config, addConfig)
 }
